requirements.txt
Flask
tensorflow
Pillow
Werkzeug
flask-cors
gunicorn
